Improved some numerical accuracy w.r.t. energy calibration. Moved to using doubles in energy calibration calculations where it makes sense (lower channel energies still stored as floats, as are coefficients and deviation pairs, but intermediate computations now use doubles). Tightened up accuracy required for answer returned by correction_due_to_dev_pairs(...).

+++ SpecUtils/CubicSpline.h
@@ -86,7 +86,7 @@ namespace SpecUtils
    respecitvely (this is non-standard behaviour for cubic splines, but how
    non-linear deviation pairs are defined).
    */
-  float eval_cubic_spline( const float x,
+  double eval_cubic_spline( const double x,
                            const std::vector<CubicSplineNode> &nodes );
